Added modifications so application can be more modular. Added final styling. Added final features.
