+++ locallibrary/catalog/templates/catalog/bookinstance_list_borrowed_user.html
@@ -3,17 +3,25 @@
 {% block content %}
     <h1>Borrowed books</h1>
 
-    {% if bookinstance_list %}
-    <ul>
+    {% for bookinst in page_obj.object_list %}
+    <hr/>
+    <li class="{% if bookinst.is_overdue %}text-danger{% endif %}">
+      <a href="{% url 'book-detail' bookinst.book.pk %}">{{ bookinst.book.title }}</a> ({{ bookinst.due_back }})
+    </li>
 
-      {% for bookinst in bookinstance_list %}
-      <li class="{% if bookinst.is_overdue %}text-danger{% endif %}">
-        <a href="{% url 'book-detail' bookinst.book.pk %}">{{ bookinst.book.title }}</a> ({{ bookinst.due_back }})
-      </li>
-      {% endfor %}
-    </ul>
-
-    {% else %}
-      <p>There are no books borrowed.</p>
-    {% endif %}
+   {% endfor %}
+   <hr/>
+<center>
+  {%if page_obj.has_previous %} {# whether the previous page exists #}
+      <a href="?page={{page_obj.previous_page_number}}"><</a> {# link to the prev page #}
+  {% endif %}
+  
+  {% if page_obj.has_other_pages %}
+  <span>{{page_obj.number}}</span> {# the current page number #}
+  {% endif %}
+ 
+  {%if page_obj.has_next %} {# whether the next page exists #}
+      <a href="?page={{page_obj.next_page_number}}">></a> {# link to the next page #}
+  {% endif %}
+</center>
 {% endblock %}
