--- locallibrary/catalog/templates/catalog/views.py
@@ -1,364 +1,0 @@
-from django.shortcuts import render
-
-# Create your views here.
-from models import Book, Author, BookInstance, Genre
-
-def index(request):
-    """View function for home page of site."""
-
-    # Generate counts of some of the main objects
-    num_books = Book.objects.all().count()
-    num_instances = BookInstance.objects.all().count()
-
-    # Available books (status = 'a')
-    num_instances_available = BookInstance.objects.filter(status__exact='a').count()
-
-    # The 'all()' is implied by default.
-    num_authors = Author.objects.count()
-
-    num_genres = Genre.objects.count()
-
-    num_books_Maze = Book.objects.filter(title__contains = 'Maze').count()
-
-    # Number of visits to this view, as counted in the session variable.
-    num_visits = request.session.get('num_visits', 0)
-    request.session['num_visits'] = num_visits + 1
-
-    context = {
-        'num_books': num_books,
-        'num_instances': num_instances,
-        'num_instances_available': num_instances_available,
-        'num_authors': num_authors,
-        'num_genres': num_genres,
-        'num_books_Maze': num_books_Maze,
-        'num_visits': num_visits,
-    }
-
-    # Render the HTML template index.html with the data in the context variable
-    return render(request, 'index.html', context=context)
-
-from django.views import generic
-
-def BookListView(request):
-    books = Book.objects.all()
-    p = Paginator(books, 10)
-    page_number = request.GET.get('page')
-    try:
-        page_obj = p.get_page(page_number)  # returns the desired page object
-    except PageNotAnInteger:
-        # if page_number is not an integer then assign the first page
-        page_obj = p.page(1)
-    except EmptyPage:
-        # if page is empty then return last page
-        page_obj = p.page(p.num_pages)
-    context = {'page_obj': page_obj}
-    # sending the page object to index.html
-    return render(request, 'book_list.html', context)
-class BookDetailView(generic.DetailView):
-    model = Book
-
-def AuthorListView(request):
-    authors = Author.objects.all()
-    p = Paginator(authors, 10)
-    page_number = request.GET.get('page')
-    try:
-        page_obj = p.get_page(page_number)  # returns the desired page object
-    except PageNotAnInteger:
-        # if page_number is not an integer then assign the first page
-        page_obj = p.page(1)
-    except EmptyPage:
-        # if page is empty then return last page
-        page_obj = p.page(p.num_pages)
-    context = {'page_obj': page_obj}
-    # sending the page object to index.html
-    return render(request, 'author_list.html', context)
-
-class AuthorDetailView(generic.DetailView):
-    model = Author
-    
-from django.contrib.auth.mixins import LoginRequiredMixin
-
-class LoanedBooksByUserListView(LoginRequiredMixin,generic.ListView):
-    """Generic class-based view listing books on loan to current user."""
-    model = BookInstance
-    template_name = 'catalog/bookinstance_list_borrowed_user.html'
-    paginate_by = 10
-
-    def get_queryset(self):
-        return (
-            BookInstance.objects.filter(borrower=self.request.user)
-            .filter(status__exact='o')
-            .order_by('due_back')
-        )
-    
-from django.contrib.auth.mixins import PermissionRequiredMixin
-
-class AllBorrowedBooksView(PermissionRequiredMixin, generic.ListView):
-    permission_required = 'catalog.can_mark_returned'
-    model = BookInstance
-    template_name = 'catalog/bookinstance_list_all_borrowed_books.html'
-    paginate_by = 10
-
-    def get_queryset(self):
-        return (
-            BookInstance.objects.filter(status__exact='o')
-            .order_by('due_back')
-        )
-
-
-import datetime
-
-from django.contrib.auth.decorators import login_required, permission_required
-from django.shortcuts import get_object_or_404
-from django.http import HttpResponseRedirect
-from django.urls import reverse
-
-from catalog.forms import RenewBookForm
-
-@login_required
-@permission_required('catalog.can_mark_returned', raise_exception=True)
-def renew_book_librarian(request, pk):
-    """View function for renewing a specific BookInstance by librarian."""
-    book_instance = get_object_or_404(BookInstance, pk=pk)
-
-    # If this is a POST request then process the Form data
-    if request.method == 'POST':
-
-        # Create a form instance and populate it with data from the request (binding):
-        form = RenewBookForm(request.POST)
-
-        # Check if the form is valid:
-        if form.is_valid():
-            # process the data in form.cleaned_data as required (here we just write it to the model due_back field)
-            book_instance.due_back = form.cleaned_data['renewal_date']
-            book_instance.save()
-
-            # redirect to a new URL:
-            return HttpResponseRedirect(reverse('all-borrowed'))
-
-    # If this is a GET (or any other method) create the default form.
-    else:
-        proposed_renewal_date = datetime.date.today() + datetime.timedelta(weeks=3)
-        form = RenewBookForm(initial={'renewal_date': proposed_renewal_date})
-
-    context = {
-        'form': form,
-        'book_instance': book_instance,
-    }
-
-    return render(request, 'catalog/book_renew_librarian.html', context)
-
-from django.views.generic.edit import CreateView, UpdateView, DeleteView
-from django.urls import reverse_lazy
-from models import Author
-
-class AuthorCreate(PermissionRequiredMixin, CreateView):
-    model = Author
-    fields = ['first_name', 'last_name', 'date_of_birth', 'date_of_death']
-    initial = {'date_of_death': '11/11/2023'}
-    permission_required = 'catalog.add_author'
-
-class AuthorUpdate(PermissionRequiredMixin, UpdateView):
-    model = Author
-    # Not recommended (potential security issue if more fields added)
-    fields = '__all__'
-    permission_required = 'catalog.change_author'
-
-class AuthorDelete(PermissionRequiredMixin, DeleteView):
-    model = Author
-    success_url = reverse_lazy('authors')
-    permission_required = 'catalog.delete_author'
-
-    def form_valid(self, form):
-        try:
-            self.object.delete()
-            return HttpResponseRedirect(self.success_url)
-        except Exception as e:
-            return HttpResponseRedirect(
-                reverse("author-delete", kwargs={"pk": self.object.pk})
-            )
-
-
-from models import Book
-
-class BookCreate(PermissionRequiredMixin, CreateView):
-    model = Book
-    fields = ['title', 'author', 'summary', 'isbn', 'genre', 'language']
-    permission_required = 'catalog.add_book'
-
-class BookUpdate(PermissionRequiredMixin, UpdateView):
-    model = Book
-    # Not recommended (potential security issue if more fields added)
-    fields = '__all__'
-    permission_required = 'catalog.change_book'
-
-class BookDelete(PermissionRequiredMixin, DeleteView):
-    model = Book
-    success_url = reverse_lazy('books')
-    permission_required = 'catalog.delete_book'
-
-    def form_valid(self, form):
-        try:
-            self.object.delete()
-            return HttpResponseRedirect(self.success_url)
-        except Exception as e:
-            return HttpResponseRedirect(
-                reverse("book-delete", kwargs={"pk": self.object.pk})
-            )
-        
-
-#New Features
-from models import Genre        
-
-def GenreListView(request):
-    genres = Genre.objects.all()
-    p = Paginator(genres, 10)
-    page_number = request.GET.get('page')
-    try:
-        page_obj = p.get_page(page_number)  # returns the desired page object
-    except PageNotAnInteger:
-        # if page_number is not an integer then assign the first page
-        page_obj = p.page(1)
-    except EmptyPage:
-        # if page is empty then return last page
-        page_obj = p.page(p.num_pages)
-    context = {'page_obj': page_obj}
-    # sending the page object to index.html
-    return render(request, 'genre_list.html', context)
-
-class GenreDetailView(generic.DetailView):
-    model = Genre
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        genre = self.get_object
-        context['books'] = Book.objects.filter(genre=self.object)
-        return context
-    
-class GenreCreate(PermissionRequiredMixin, CreateView):
-    model = Genre
-    fields = ['name']
-    permission_required = 'catalog.add_genre'
-
-class GenreUpdate(PermissionRequiredMixin, UpdateView):
-    model = Genre
-    # Not recommended (potential security issue if more fields added)
-    fields = '__all__'
-    permission_required = 'catalog.change_genre'
-
-class GenreDelete(PermissionRequiredMixin, DeleteView):
-    model = Genre
-    success_url = reverse_lazy('genres')
-    permission_required = 'catalog.delete_genre'
-
-    def form_valid(self, form):
-        try:
-            self.object.delete()
-            return HttpResponseRedirect(self.success_url)
-        except Exception as e:
-            return HttpResponseRedirect(
-                reverse("genre-delete", kwargs={"pk": self.object.pk})
-            )
-        
-from models import Language        
-
-def LanguageListView(request):
-    languages = Language.objects.all()
-    p = Paginator(languages, 10)
-    page_number = request.GET.get('page')
-    try:
-        page_obj = p.get_page(page_number)  # returns the desired page object
-    except PageNotAnInteger:
-        # if page_number is not an integer then assign the first page
-        page_obj = p.page(1)
-    except EmptyPage:
-        # if page is empty then return last page
-        page_obj = p.page(p.num_pages)
-    context = {'page_obj': page_obj}
-    # sending the page object to index.html
-    return render(request, 'language_list.html', context)
-
-class LanguageDetailView(generic.DetailView):
-    model = Language
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        language = self.get_object
-        context['books'] = Book.objects.filter(language=self.object)
-        return context
-    
-class LanguageCreate(PermissionRequiredMixin, CreateView):
-    model = Language
-    fields = ['name']
-    permission_required = 'catalog.add_language'
-
-class LanguageUpdate(PermissionRequiredMixin, UpdateView):
-    model = Language
-    # Not recommended (potential security issue if more fields added)
-    fields = '__all__'
-    permission_required = 'catalog.change_language'
-
-class LanguageDelete(PermissionRequiredMixin, DeleteView):
-    model = Language
-    success_url = reverse_lazy('languages')
-    permission_required = 'catalog.delete_language'
-
-    def form_valid(self, form):
-        try:
-            self.object.delete()
-            return HttpResponseRedirect(self.success_url)
-        except Exception as e:
-            return HttpResponseRedirect(
-                reverse("language-delete", kwargs={"pk": self.object.pk})
-            )
-        
-from models import BookInstance
-from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger       
-
-def BookInstanceListView(request):
-    books = BookInstance.objects.all()
-    p = Paginator(books, 10)
-    page_number = request.GET.get('page')
-    try:
-        page_obj = p.get_page(page_number)  # returns the desired page object
-    except PageNotAnInteger:
-        # if page_number is not an integer then assign the first page
-        page_obj = p.page(1)
-    except EmptyPage:
-        # if page is empty then return last page
-        page_obj = p.page(p.num_pages)
-    context = {'page_obj': page_obj}
-    # sending the page object to index.html
-    return render(request, 'bookinstance_list.html', context)
-
-class BookInstanceDetailView(generic.DetailView):
-    model = BookInstance
-    
-class BookInstanceCreate(PermissionRequiredMixin, CreateView):
-    model = BookInstance
-    fields = ['book', 'imprint', 'due_back', 'borrower', 'status']
-    permission_required = 'catalog.add_bookinstance'
-
-class BookInstanceUpdate(PermissionRequiredMixin, UpdateView):
-    model = BookInstance
-    # Not recommended (potential security issue if more fields added)
-    fields = '__all__'
-    permission_required = 'catalog.change_bookinstance'
-
-class BookInstanceDelete(PermissionRequiredMixin, DeleteView):
-    model = BookInstance
-    success_url = reverse_lazy('bookinstances')
-    permission_required = 'catalog.delete_bookinstance'
-
-    def form_valid(self, form):
-        try:
-            self.object.delete()
-            return HttpResponseRedirect(self.success_url)
-        except Exception as e:
-            return HttpResponseRedirect(
-                reverse("bookInstance-delete", kwargs={"pk": self.object.pk})
-            )
-        
-
-class DetailView(generic.DetailView):
-    model = generic.DetailView.model
