+++ locallibrary/catalog/urls.py
@@ -3,12 +3,12 @@ from . import views
 
 urlpatterns = [
     path('', views.index, name='index'),
-    path('books/', views.BookListView.as_view(), name='books'),
+    path('books/', views.BookListView, name='books'),
     path('book/<int:pk>', views.BookDetailView.as_view(), name='book-detail'),
-    path('authors/', views.AuthorListView.as_view(), name='authors'),
+    path('authors/', views.AuthorListView, name='authors'),
     path('author/<int:pk>', views.AuthorDetailView.as_view(), name='author-detail'),
-    path('mybooks/', views.LoanedBooksByUserListView.as_view(), name='my-borrowed'),
-    path('allborrowed/', views.AllBorrowedBooksView.as_view(), name='all-borrowed'),
+    path('mybooks/', views.LoanedBooksByUserListView, name='my-borrowed'),
+    path('allborrowed/', views.AllBorrowedBooksView, name='all-borrowed'),
     path('book/<uuid:pk>/renew/', views.renew_book_librarian, name='renew-book-librarian'),
     path('author/create/', views.AuthorCreate.as_view(), name='author-create'),
     path('author/<int:pk>/update/', views.AuthorUpdate.as_view(), name='author-update'),
@@ -18,19 +18,19 @@ urlpatterns = [
     path('book/<int:pk>/delete/', views.BookDelete.as_view(), name='book-delete'),
 
     #New urls
-    path('genres/', views.GenreListView.as_view(), name='genres'),
+    path('genres/', views.GenreListView, name='genres'),
     path('genre/<int:pk>', views.GenreDetailView.as_view(), name='genre-detail'),
     path('genre/create/', views.GenreCreate.as_view(), name='genre-create'),
     path('genre/<int:pk>/update/', views.GenreUpdate.as_view(), name='genre-update'),
     path('genre/<int:pk>/delete/', views.GenreDelete.as_view(), name='genre-delete'),
 
-    path('languages/', views.LanguageListView.as_view(), name='languages'),
+    path('languages/', views.LanguageListView, name='languages'),
     path('language/<int:pk>', views.LanguageDetailView.as_view(), name='language-detail'),
     path('language/create/', views.LanguageCreate.as_view(), name='language-create'),
     path('language/<int:pk>/update/', views.LanguageUpdate.as_view(), name='language-update'),
     path('language/<int:pk>/delete/', views.LanguageDelete.as_view(), name='language-delete'),
 
-    path('bookinstances/', views.BookInstanceListView.as_view(), name='bookinstances'),
+    path('bookinstances/', views.BookInstanceListView, name='bookinstances'),
     path('bookinstance/<uuid:pk>', views.BookInstanceDetailView.as_view(), name='bookinstance-detail'),
     path('bookinstance/create/', views.BookInstanceCreate.as_view(), name='bookinstance-create'),
     path('bookinstance/<uuid:pk>/update/', views.BookInstanceUpdate.as_view(), name='bookinstance-update'),
